Fix backwards pagination
Change the order of the query to make sure the right result set
is returned when using backwards pagination.

At the end, using Array.reverse() when there's a `last` argument
ensures the order is kept the same, and only the direction is changed.
This way, the backwards pagination works, and it will still be on par
with [1]

[1] https://relay.dev/graphql/connections.htm